requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
build==1.4.4
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
cohere==6.1.0
colorama==0.4.6
cryptography==46.0.7
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
fastavro==1.12.2
filelock==3.29.0
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.47
google-auth==2.49.2
google-genai==1.73.1
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.12.0
idna==3.13
importlib_metadata==8.7.1
importlib_resources==7.1.0
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.2
langchain-google-genai==4.2.2
langchain-openai==1.2.0
langchain-protocol==0.0.11
langchain-text-splitters==1.1.2
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.36
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.20.0
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.25.0
openai==2.32.0
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
propcache==0.4.1
protobuf==6.33.6
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.14.0
pydantic_core==2.41.5
pydeck==0.9.2
Pygments==2.20.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
streamlit==1.56.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.2
types-requests==2.33.0.20260408
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
